requirements.txt
numpy
opencv-python
Pillow
pdf2image
pytesseract
scipy
scikit-image
PyYAML
jpegio
torch
torchvision